requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0

# Real-time and WebSocket support
websockets>=11.0.0
asyncio>=3.4.3
aiohttp>=3.8.0

# Quantum computing
qiskit>=0.45.0
qiskit-aer>=0.12.0
qiskit-algorithms>=0.2.0

# Machine Learning
torch>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# Network analysis and real-time monitoring
scapy>=2.5.0
psutil>=5.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.18.0

# Web framework
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
sqlite3
sqlalchemy>=2.0.0

# Utilities
pyyaml>=6.0
python-dateutil>=2.8.0
colorama>=0.4.6
tqdm>=4.65.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0

# Real-time data processing
redis>=4.6.0
celery>=5.3.0

# Optional quantum backends
# qiskit-ibm-runtime>=0.11.0  # For IBM Quantum
# amazon-braket-sdk>=1.60.0   # For AWS Braket
